requirements.txt
pyautogui
pytesseract
opencv-python
pandas
Pillow
keyboard
mouse
